# backend/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api GerencIA backend.",
        "title": "Api GerencIA backend",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Swagger API Support",
            "url": "http://www.swagger.io/support",
            "email": "linhar3s77@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/card/create": {
            "post": {
                "description": "Create a new card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Create a new card",
                "parameters": [
                    {
                        "description": "Card Data",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/dtos.CardDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {}
                }
            }
        },
        "/card/delete/{id}": {
            "delete": {
                "description": "Delete a card by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Delete a card by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted the card",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/card/list": {
            "get": {
                "description": "Get all cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Get all cards",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Query parameters",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Query parameters",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Query parameters",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/card/update": {
            "patch": {
                "description": "Update a card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Update a card",
                "parameters": [
                    {
                        "description": "Card Data",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/dtos.CardDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {}
                }
            }
        }
    },
    "definitions": {
        "dtos.CardDTO": {
            "type": "object",
            "required": [
                "limite",
                "nome"
            ],
            "properties": {
                "atualizado_em": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "limite": {
                    "type": "number"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "has_next": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}